move all builtin functions into `builtins.go` file
