--- builtin_type.go
@@ -1,9 +1,0 @@
-package tengo
-
-func builtinTypeName(_ Interop, args ...Object) (Object, error) {
-	if len(args) != 1 {
-		return nil, ErrWrongNumArguments
-	}
-
-	return &String{Value: args[0].TypeName()}, nil
-}
