--- builtin_type_checks.go
@@ -1,194 +1,0 @@
-package tengo
-
-func builtinIsString(_ Interop, args ...Object) (Object, error) {
-	if len(args) != 1 {
-		return nil, ErrWrongNumArguments
-	}
-
-	if _, ok := args[0].(*String); ok {
-		return TrueValue, nil
-	}
-
-	return FalseValue, nil
-}
-
-func builtinIsInt(_ Interop, args ...Object) (Object, error) {
-	if len(args) != 1 {
-		return nil, ErrWrongNumArguments
-	}
-
-	if _, ok := args[0].(*Int); ok {
-		return TrueValue, nil
-	}
-
-	return FalseValue, nil
-}
-
-func builtinIsFloat(_ Interop, args ...Object) (Object, error) {
-	if len(args) != 1 {
-		return nil, ErrWrongNumArguments
-	}
-
-	if _, ok := args[0].(*Float); ok {
-		return TrueValue, nil
-	}
-
-	return FalseValue, nil
-}
-
-func builtinIsBool(_ Interop, args ...Object) (Object, error) {
-	if len(args) != 1 {
-		return nil, ErrWrongNumArguments
-	}
-
-	if _, ok := args[0].(*Bool); ok {
-		return TrueValue, nil
-	}
-
-	return FalseValue, nil
-}
-
-func builtinIsChar(_ Interop, args ...Object) (Object, error) {
-	if len(args) != 1 {
-		return nil, ErrWrongNumArguments
-	}
-
-	if _, ok := args[0].(*Char); ok {
-		return TrueValue, nil
-	}
-
-	return FalseValue, nil
-}
-
-func builtinIsBytes(_ Interop, args ...Object) (Object, error) {
-	if len(args) != 1 {
-		return nil, ErrWrongNumArguments
-	}
-
-	if _, ok := args[0].(*Bytes); ok {
-		return TrueValue, nil
-	}
-
-	return FalseValue, nil
-}
-
-func builtinIsArray(_ Interop, args ...Object) (Object, error) {
-	if len(args) != 1 {
-		return nil, ErrWrongNumArguments
-	}
-
-	if _, ok := args[0].(*Array); ok {
-		return TrueValue, nil
-	}
-
-	return FalseValue, nil
-}
-
-func builtinIsImmutableArray(_ Interop, args ...Object) (Object, error) {
-	if len(args) != 1 {
-		return nil, ErrWrongNumArguments
-	}
-
-	if _, ok := args[0].(*ImmutableArray); ok {
-		return TrueValue, nil
-	}
-
-	return FalseValue, nil
-}
-
-func builtinIsMap(_ Interop, args ...Object) (Object, error) {
-	if len(args) != 1 {
-		return nil, ErrWrongNumArguments
-	}
-
-	if _, ok := args[0].(*Map); ok {
-		return TrueValue, nil
-	}
-
-	return FalseValue, nil
-}
-
-func builtinIsImmutableMap(_ Interop, args ...Object) (Object, error) {
-	if len(args) != 1 {
-		return nil, ErrWrongNumArguments
-	}
-
-	if _, ok := args[0].(*ImmutableMap); ok {
-		return TrueValue, nil
-	}
-
-	return FalseValue, nil
-}
-
-func builtinIsTime(_ Interop, args ...Object) (Object, error) {
-	if len(args) != 1 {
-		return nil, ErrWrongNumArguments
-	}
-
-	if _, ok := args[0].(*Time); ok {
-		return TrueValue, nil
-	}
-
-	return FalseValue, nil
-}
-
-func builtinIsError(_ Interop, args ...Object) (Object, error) {
-	if len(args) != 1 {
-		return nil, ErrWrongNumArguments
-	}
-
-	if _, ok := args[0].(*Error); ok {
-		return TrueValue, nil
-	}
-
-	return FalseValue, nil
-}
-
-func builtinIsUndefined(_ Interop, args ...Object) (Object, error) {
-	if len(args) != 1 {
-		return nil, ErrWrongNumArguments
-	}
-
-	if args[0] == UndefinedValue {
-		return TrueValue, nil
-	}
-
-	return FalseValue, nil
-}
-
-func builtinIsFunction(_ Interop, args ...Object) (Object, error) {
-	if len(args) != 1 {
-		return nil, ErrWrongNumArguments
-	}
-
-	switch args[0].(type) {
-	case *CompiledFunction:
-		return TrueValue, nil
-	}
-
-	return FalseValue, nil
-}
-
-func builtinIsCallable(_ Interop, args ...Object) (Object, error) {
-	if len(args) != 1 {
-		return nil, ErrWrongNumArguments
-	}
-
-	if args[0].CanCall() {
-		return TrueValue, nil
-	}
-
-	return FalseValue, nil
-}
-
-func builtinIsIterable(_ Interop, args ...Object) (Object, error) {
-	if len(args) != 1 {
-		return nil, ErrWrongNumArguments
-	}
-
-	if args[0].CanIterate() {
-		return TrueValue, nil
-	}
-
-	return FalseValue, nil
-}
